fix: use parsed terminal output for summary component
Instead of trying to strip terminal escape codes from the output, it
works a lot better if we just take the output directly from the terminal
buffer. The only wrinkle is that the terminal buffer isn't updated when
we get the output callback, so we have to poll it again after a short
delay.