feat(summary): Skip [Process exited] in summary
With the latest commit, the Overseer sidebar now displays the helpful
summary of the tast. But for a shell command of c++ compile commands,
the last line is always [Process extied err_code] which is not helpful
because the task exit notification already states if it was success or
failure. This PR just gets the next non_empty line from the bottom