fix: exclude dependency tasks from bundles
Suppose the case of a task with a dependency. It starts the dependency,
then runs itself. Previously, we would save both tasks in
OverseerSaveBundle. When we loaded the bundle, we would rehydrate one
copy of the dependency task, and the rehydrated master task would create
another copy of it, leading to duplicate dependency tasks.

After this change, we exclude the dependency tasks from the bundle. They
will be recreated by the master task on load.