feat: add new prompt value 'avoid' and rework 'allow' (#57)
Previously the default value of "allow" would only show the template
parameter prompt if there were any params that were not optional AND
didn't have a default value. In effect this meant that template
parameters could not separate the concept of "the user should be forced
to choose a value for this parameter" and "this is the default value for
this parameter". In particular this had a bad interaction with VS Code
tasks (see linked issue).

The new behavior for "allow" will prompt the user if any non-optional
params were not supplied to run_template, regardless of if they have a
default value. This should allow templates to have a bit more granular
control over the prompt behavior. For users that want the old behavior
of not prompting unless absolutely required, they can use the new option
of "avoid", which will avoid the prompt unless absolutely required.