refactor: wrapped templates add optional=true to default params
When wrapping templates, we are usually supplying a particular subcmd or
argument that the user will select directly in the task selector. This
means they usually don't want to change that value, so we can set
`optional = true` to not prompt them (when using the default "allow"
setting). If there are any other parameters for the task that are
`optional = false`, the user can still be prompted.