fix: make toggleterm terminal window close when task resets
Before running the `restart` action on a task with the toggleterm
strategy would make a second window appear if one was already open. This
change calls `term:close()` in the `reset` method of the toggleterm
strategy, to fix this issue.

To have access to the terminal handler in the `reset` method, I had to
save it in the object as `self.term`.