fix!: don't auto-add on_result_diagnostics to VS Code tasks (#163)
This is a breaking change if you have defined the `default_vscode`
component alias. Previously if the task had a problemMatcher, we would
_always_ add the `on_result_diagnostics` component. This doesn't really
make sense, as that component is present in the `default_vscode` alias
so it will generally be present. Auto-adding it prevents the end user
from being able to _not_ have that component attached.