npm: smarter package.json file detection
fixes #242
Some projects have package.json files in subfolders, which are not the main project package.json file but rather some submodule marker. This seems prevalent in react-native projects. See this for instance:
https://stackoverflow.com/questions/51701191/react-native-has-something-to-use-local-folders-as-package-name-what-is-it-ca
To cover that case, we search for package.json files starting from the current file folder, and pick the first matching candidate.