fix: always parse make output to find tasks (#280)
We also have a function that uses treesitter to parse the Makefile and
get the list of tasks from that, but it doesn't handle the `include`
command mentioned in the issue. The weird flakiness was because on the
first run the pcall to get the parser failed somewhere in _memoize
(which seems like a Neovim bug), so we would fall back to parsing the
output of `make` and actually produce the correct result. On subsequent
runs, the parser would be fetched correctly and so the results would be
missing the included tasks. The only advantage of the treesitter method
is that it's faster, but parsing the output of `make` is fast enough,
and it's better to be correct. We can just get rid of the treesitter
logic entirely and always rely on parsing the output of `make`.