fix(on_result_diagnostic): treat `end_lnum < 1` as uninitialized (#396)
* fix(on_result_diagnostic): treat `end_lnum < 1` as unitialized

The end of the diagnostic range was previously calculated as `item.end_lnum and item.end_lnum - 1`. Given `item = { lnum: 95, end_lnum: 0, ... }`, this calculation would span everything from the beginning up to line 95. I believe that an `end_lnum` value of zero should be treated as uninitialized (because lines are 1-indexed). Therefore, I think the item should be considered a single-line item.

* refactor(on_result_diagnostic): remove ternary operator